Splitting the widget.
