+++ UI/XPlat/flutter_template/lib/components/importer.dart
@@ -7,9 +7,12 @@ import 'package:flutter/material.dart';
 export 'package:flutter_template/configs/app_fcm.dart';
 export 'package:flutter_template/models/message_arguments.dart';
 
-export 'package:flutter_template/common/permissions.dart';
-export 'package:flutter_template/common/token_monitor.dart';
-export 'package:flutter_template/common/message_list.dart';
 export 'package:flutter_template/common/meta_card.dart';
+export 'package:flutter_template/common/permissions.dart';
+export 'package:flutter_template/common/token_checker.dart';
+export 'package:flutter_template/common/message_list.dart';
 
+export 'package:flutter_template/components/my_home_page.dart';
+export 'package:flutter_template/components/appauth_page.dart';
+export 'package:flutter_template/components/fcm_page.dart';
 export 'package:flutter_template/components/message_view.dart';
