+++ UI/XPlat/flutter_template/lib/configs/app_fcm.dart
@@ -15,4 +15,8 @@ class AppFcm {
   /// Initialize the [FlutterLocalNotificationsPlugin] package.
   /// FlutterLocalNotificationsPlugin]パッケージを初期化します。
   static FlutterLocalNotificationsPlugin? flutterLocalNotificationsPlugin;
+
+  static const vapidKey = '<YOUR_PUBLIC_VAPID_KEY_HERE>';
+  static String? token = "hoge";
+  static Stream<String>? tokenStream;
 }
